Split domain records into individual files
