+++ src/main/java/com/example/reporting/domain/Column.java
@@ -1,0 +1,8 @@
+package com.example.reporting.domain;
+
+public record Column(
+        String header,
+        String field,
+        String format,
+        String style
+) {}
