+++ src/main/java/com/example/reporting/domain/DatasetDef.java
@@ -1,0 +1,12 @@
+package com.example.reporting.domain;
+
+import java.util.List;
+import java.util.Map;
+
+public record DatasetDef(
+        String name,
+        SourceDef source,
+        List<Map<String, Object>> transforms,
+        List<Map<String, Object>> validations,
+        Boolean cache
+) {}
