+++ src/main/java/com/example/reporting/domain/OutputDef.java
@@ -1,0 +1,6 @@
+package com.example.reporting.domain;
+
+public record OutputDef(
+        String filenamePattern,
+        String storage
+) {}
