+++ src/main/java/com/example/reporting/domain/ReportDefinition.java
@@ -1,0 +1,11 @@
+package com.example.reporting.domain;
+
+import java.util.List;
+
+public record ReportDefinition(
+        String reportCode,
+        String version,
+        TemplateDef template,
+        List<DatasetDef> datasets,
+        OutputDef output
+) {}
