+++ src/main/java/com/example/reporting/domain/SheetBinding.java
@@ -1,0 +1,8 @@
+package com.example.reporting.domain;
+
+import java.util.List;
+
+public record SheetBinding(
+        String name,
+        List<Binding> bindings
+) {}
