+++ src/main/java/com/example/reporting/domain/SourceDef.java
@@ -1,0 +1,11 @@
+package com.example.reporting.domain;
+
+import java.util.Map;
+
+public record SourceDef(
+        String kind,
+        String connection,
+        String query,
+        String path,
+        Map<String, Object> options
+) {}
