+++ src/main/java/com/example/reporting/domain/TemplateDef.java
@@ -1,0 +1,8 @@
+package com.example.reporting.domain;
+
+import java.util.List;
+
+public record TemplateDef(
+        String path,
+        List<SheetBinding> sheets
+) {}
